+++ LabeSystem/src/endpoints/addStudentsClass.ts
@@ -1,0 +1,6 @@
+import {Request, Response} from "express"
+
+export const addStudentClass = async (req:Request, res: Response): Promise<void> =>{
+
+    
+}
