+++ LabeSystem/src/endpoints/addTeacherClass.ts
@@ -1,0 +1,6 @@
+import {Request, Response} from "express"
+
+export const addTeacherClass = async (req:Request, res: Response): Promise<void> =>{
+
+    
+}
