+++ LabeSystem/src/endpoints/createClass.ts
@@ -1,0 +1,6 @@
+import {Request, Response} from "express"
+
+export const createClass = async (req:Request, res: Response): Promise<void> =>{
+
+    
+}
