+++ LabeSystem/src/endpoints/createStudent.ts
@@ -1,0 +1,6 @@
+import {Request, Response} from "express"
+
+export const createStudent = async (req:Request, res: Response): Promise<void> =>{
+
+    
+}
