+++ LabeSystem/src/endpoints/createTeacher.ts
@@ -1,0 +1,6 @@
+import {Request, Response} from "express"
+
+export const createTeacher = async (req:Request, res: Response): Promise<void> =>{
+
+    
+}
