+++ LabeSystem/src/endpoints/getStudentAgeById.ts
@@ -1,0 +1,6 @@
+import {Request, Response} from "express"
+
+export const getStudentAgeById = async (req:Request, res: Response): Promise<void> =>{
+
+    
+}
