+++ LabeSystem/src/index.ts
@@ -1,31 +1,14 @@
-import express, { Express } from "express";
-import knex from "knex";
-import cors from "cors";
-import dotenv from "dotenv";
-import { AddressInfo } from "net";
+import {app} from "./app"
+import { addStudentClass } from "./endpoints/addStudentsClass";
+import { addTeacherClass } from "./endpoints/addTeacherClass";
+import { createClass } from "./endpoints/createClass";
+import { createStudent } from "./endpoints/createStudent";
+import { createTeacher } from "./endpoints/createTeacher";
+import { getStudentAgeById } from "./endpoints/getStudentAgeById";
 
-dotenv.config();
-
-export const connection = knex({
-	client: "mysql",
-	connection: {
-    host: process.env.DB_HOST,
-    port: 3306,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASS,
-    database: process.env.DB_NAME
-  }
-});
-
-const app: Express = express();
-app.use(express.json());
-app.use(cors());
-
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-       const address = server.address() as AddressInfo;
-       console.log(`Server is running in http://localhost: ${address.port}`);
-    } else {
-       console.error(`Failure upon starting server.`);
-    }
-});
+app.get("/student/:id", getStudentAgeById)
+app.post("/class", createClass)
+app.post("/teacher",createTeacher)
+app.post("/student",createStudent)
+app.patch("/student/:id",addStudentClass)
+app.patch("/teacher/:id", addTeacherClass)
